# angular-app/src/openapi/openapi3.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Group Service API",
    "description": "API for managing educational groups secured with Keycloak",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8083",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/posts": {
      "get": {
        "tags": [
          "Post"
        ],
        "operationId": "getAllPosts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelPost"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Post"
        ],
        "operationId": "createPost",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/{postId}/comments": {
      "get": {
        "tags": [
          "Post"
        ],
        "operationId": "getCommentsWithReplies",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "format": "int64"
                      },
                      "content": {
                        "type": "string"
                      },
                      "authorId": {
                        "type": "string"
                      },
                      "postId": {
                        "type": "integer",
                        "format": "int64"
                      },
                      "parentCommentId": {
                        "type": "integer",
                        "format": "int64"
                      },
                      "createdDate": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "replies": {
                        "type": "array",
                        "items": {

                        }
                      },
                      "reply": {
                        "type": "boolean"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Post"
        ],
        "operationId": "addComment",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "content",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parentCommentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/{post-id}/like": {
      "post": {
        "tags": [
          "Post"
        ],
        "operationId": "toggleLike",
        "parameters": [
          {
            "name": "post-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LikeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/{post-id}/bookmark": {
      "post": {
        "tags": [
          "Post"
        ],
        "operationId": "toggleBookmark",
        "parameters": [
          {
            "name": "post-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/posts/{post-id}/status": {
      "patch": {
        "tags": [
          "Post"
        ],
        "operationId": "updatePostStatus",
        "parameters": [
          {
            "name": "post-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/{post-id}/pin": {
      "patch": {
        "tags": [
          "Post"
        ],
        "operationId": "pinPost",
        "parameters": [
          {
            "name": "post-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "pinned",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/{postId}/comments/paginated": {
      "get": {
        "tags": [
          "Post"
        ],
        "operationId": "getCommentsPaginated",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelCommentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/{post-id}/liked": {
      "get": {
        "tags": [
          "Post"
        ],
        "operationId": "isPostLikedByUser",
        "parameters": [
          {
            "name": "post-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/{post-id}/bookmarked": {
      "get": {
        "tags": [
          "Post"
        ],
        "operationId": "isPostBookmarkedByUser",
        "parameters": [
          {
            "name": "post-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/trending": {
      "get": {
        "tags": [
          "Post"
        ],
        "operationId": "getTrendingPosts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelPost"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/search": {
      "get": {
        "tags": [
          "Post"
        ],
        "operationId": "searchPosts",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelPost"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/my-posts": {
      "get": {
        "tags": [
          "Post"
        ],
        "operationId": "getMyPosts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelPost"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/my-pending": {
      "get": {
        "tags": [
          "Post"
        ],
        "operationId": "getMyPendingPosts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelPost"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/bookmarks": {
      "get": {
        "tags": [
          "Post"
        ],
        "operationId": "getBookmarkedPosts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelPost"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/posts/{post-id}": {
      "delete": {
        "tags": [
          "Post"
        ],
        "operationId": "deletePost",
        "parameters": [
          {
            "name": "post-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreatePostRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 0
          },
          "imageUrl": {
            "type": "string"
          },
          "imageFile": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "content"
        ]
      },
      "Bookmark": {
        "type": "object",
        "properties": {
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "string"
          }
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "authorId": {
            "type": "string"
          },
          "approved": {
            "type": "boolean"
          }
        }
      },
      "Like": {
        "type": "object",
        "properties": {
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "string"
          }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "authorId": {
            "type": "string"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Like"
            }
          },
          "bookmarks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Bookmark"
            }
          },
          "likeCount": {
            "type": "integer",
            "format": "int64"
          },
          "commentCount": {
            "type": "integer",
            "format": "int64"
          },
          "bookmarkCount": {
            "type": "integer",
            "format": "int64"
          },
          "pinned": {
            "type": "boolean"
          },
          "approved": {
            "type": "boolean"
          },
          "summary": {
            "type": "string"
          },
          "pending": {
            "type": "boolean"
          },
          "rejected": {
            "type": "boolean"
          }
        }
      },
      "CommentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "authorId": {
            "type": "string"
          },
          "postId": {
            "type": "integer",
            "format": "int64"
          },
          "parentCommentId": {
            "type": "integer",
            "format": "int64"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "replies": {
            "type": "array",
            "items": {

            }
          },
          "reply": {
            "type": "boolean"
          }
        }
      },
      "LikeResponse": {
        "type": "object",
        "properties": {
          "like": {
            "$ref": "#/components/schemas/Like"
          },
          "liked": {
            "type": "boolean"
          },
          "action": {
            "type": "string"
          },
          "likeCount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BookmarkResult": {
        "type": "object",
        "properties": {
          "bookmark": {
            "$ref": "#/components/schemas/Bookmark"
          },
          "bookmarked": {
            "type": "boolean"
          },
          "bookmarkCount": {
            "type": "integer",
            "format": "int64"
          },
          "action": {
            "type": "string"
          }
        }
      },
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PagedModelPost": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
      "PagedModelCommentResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentResponse"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Provide the JWT token obtained from Keycloak: 'Bearer eyJ...'",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}